Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -3,9 +3,9 @@
 int main()
 {
     int a,b,c;
-    float avg;
+    double avg;
     scanf("%d%d%d",&a,&b,&c);
     avg=(a+b+c)/3;
-    printf("Average: %.2f",avg);
+    printf("Average: %.2lf",avg);
 
 }
